pyproject.toml
[project]
name = "icb-qa-automation"
version = "0.1.0"
description = ""
authors = [
    {name = "Youssef chakir",email = "76889503+Yousssefch@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.38.0,<5.0.0)",
    "pytest-html (>=4.1.1,<5.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "icb_qa_automation", from = "src"}]

[tool.poetry.scripts]
main = "icb_qa_automation.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
